+++ CameraHal/CameraIspAdapter.h
@@ -62,7 +62,7 @@ private:
     //talk to driver
     virtual int cameraCreate(int cameraId);
     virtual int cameraDestroy();
-    virtual int adapterReturnFrame(int index,int cmd);
+    virtual int adapterReturnFrame(long index,int cmd);
 
 
     //for isp
